ML Lab/lab3/requirements.txt
matplotlib>=3.2.2
numpy==1.26.2
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
--find-links https://download.pytorch.org/whl/cu121
torch==2.2.2
torchvision==0.17.2
tqdm>=4.41.0
protobuf
tensorboardX
pandas>=1.1.4
seaborn>=0.11.0
